requirements.txt
dash
plotly
paho-mqtt
gunicorn